+++ th2_data_services/adapter.py
@@ -5,7 +5,7 @@ from typing import Union, List
 def change_pipeline_message(record: dict) -> Union[List[dict], dict]:
     msg_type = record.get("messageType")
     if not msg_type:
-        warn("Please note, some messages don't have a messageType field.", stacklevel=3)
+        warn("Please note, some messages don't have a messageType field. Perhaps a codec doesn't decode some messages.", stacklevel=3)
         return record
 
     if "/" not in msg_type: